Refactored date range calculation into utility file
