+++ assets/wire/actions.js
@@ -175,6 +175,12 @@ function search(state, next) {
     const newsOnly = !!get(state, 'wire.newsOnly');
     const context = get(state, 'context', 'wire');
 
+    let created_to = createdFilter.to;
+    if (createdFilter.from && createdFilter.from.indexOf('now') >= 0) {
+        created_to = createdFilter.from;
+    }
+
+
     const params = {
         q: state.query,
         bookmarks: state.bookmarks && state.user,
@@ -182,7 +188,7 @@ function search(state, next) {
         filter: !isEmpty(activeFilter) && JSON.stringify(activeFilter),
         from: next ? state.items.length : 0,
         created_from: createdFilter.from,
-        created_to: createdFilter.to,
+        created_to,
         timezone_offset: getTimezoneOffset(),
         newsOnly
     };
